Split off time formats to their own file.
