+++ astropy/time/__init__.py
@@ -1,2 +1,3 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
+from .formats import *
 from .core import *
